CASSSIDECAR-214: Speed up ConnectedClientStatsHandlerIntegrationTest
Patch by Francisco Guerrero; reviewed by TBD for CASSSIDECAR-214
